requirements.txt
python-telegram-bot==5.3.1
requests==2.2.1